refactor: created new types.ts in server folder for server specific types, separated from shared types
