+++ src/features/knowledgebase/actions/embed.ts
@@ -1,5 +1,8 @@
 import { embed as embedFromAdapter } from "../../../server/llm/providers/gemini";
-import { EmbedRequest, EmbedResponse } from "../../shared/lib/types";
+import {
+  EmbedRequest,
+  EmbedResponse,
+} from "../../../server/llm/types";
 
 // Generate embedding
 export async function embed({ text }: EmbedRequest): Promise<EmbedResponse> {
